Implemented sms service; Added sms helper services; Changed dapper with EF core; Changed entities' configuration; Added NLog; Added helper extensions;

+++ src/Core/MessageSender.Domain/Contracts/ISmsRepository.cs
@@ -4,6 +4,6 @@ namespace MessageSender.Domain.Contracts;
 
 public interface ISmsRepository
 {
-    Task<int> InsertSmsAsync(Guid clientId, string phoneNumber, CancellationToken cancellationToken = default);
+    Task<long> InsertSmsAsync(Guid clientId, string phoneNumber, CancellationToken cancellationToken = default);
     Task UpdateSmsContentAsync(int smsId, string content, CancellationToken cancellationToken = default);
 }
